Final_project/requirements.txt
pytest
pytest-selenium
termcolor
allure-python-commons